requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3 

python-dotenv
num2words==0.5.12
bcrypt==4.0.1
qrcode[pil]==7.4.2
selenium==4.23.1
pytest==8.3.2
pytest-html==3.2.0
pytesseract==0.3.13
Pillow==11.3.0
# opencv-python==4.12.0.88  # Optional - will use PIL fallback if not available

# PostgreSQL dependencies for Railway deployment
psycopg2-binary==2.9.7
psycopg2-pool==1.1

# Security packages
Flask-WTF==1.1.1
Flask-Limiter==3.5.0
marshmallow==3.20.1
cryptography==41.0.7

# Production WSGI server
gunicorn==21.2.0
